backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.30
alembic==1.12.1
psycopg2-binary==2.9.9; platform_system != "Darwin" or python_version < "3.13"
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
openai==1.3.5
sse-starlette==1.8.2
httpx==0.28.1

